requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docopt==0.6.2
idna==3.2
Markdown==3.3.4
Pillow==8.3.1
pipreqs==0.4.10
psycopg2==2.9.1
PyJWT==2.0.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
yarg==0.1.9
